+++ src/engine/info-plugin.ts
@@ -13,30 +13,6 @@ export interface EngineInfo {
 
 export const engineInfo = Symbol()
 
-/**
- * Extracts the title from the higher hierarchy heading in the Markdown tokens.
- * @param tokens - The list of Markdown tokens.
- * @returns The extracted title or undefined if no heading is found.
- */
-function extractTitleFromFirstHeading(tokens: any[]): string | undefined {
-  let bestHeading: { level: number; content: string } | undefined
-  for (let i = 0; i < tokens.length; i++) {
-    const token = tokens[i]
-    if (token.type === 'heading_open') {
-      const level = parseInt(token.tag.slice(1), 10)
-      const headingContentToken = tokens[i + 1]
-      if (
-        headingContentToken &&
-        headingContentToken.type === 'inline' &&
-        (!bestHeading || level < bestHeading.level)
-      ) {
-        bestHeading = { level, content: headingContentToken.content.trim() }
-      }
-    }
-  }
-  return bestHeading?.content
-}
-
 export default function infoPlugin(md: any) {
   const { marpit } = md
 
@@ -46,9 +22,6 @@ export default function infoPlugin(md: any) {
     const { themeSet, lastGlobalDirectives } = marpit
     const globalDirectives = lastGlobalDirectives || {}
     const theme = globalDirectives.theme || (themeSet.default || {}).name
-    const title =
-      globalDirectives.marpCLITitle ??
-      extractTitleFromFirstHeading(state.tokens)
 
     const info: EngineInfo = {
       theme,
@@ -57,7 +30,7 @@ export default function infoPlugin(md: any) {
       image: globalDirectives.marpCLIImage,
       keywords: globalDirectives.marpCLIKeywords,
       lang: globalDirectives.lang || marpit.options.lang,
-      title,
+      title: globalDirectives.marpCLITitle,
       url: globalDirectives.marpCLIURL,
       size: {
         height: themeSet.getThemeProp(theme, 'heightPixel'),