+++ src/engine/meta-plugin.ts
@@ -1,6 +1,10 @@
 import { URL } from 'node:url'
 import { Marpit } from '@marp-team/marpit'
+import type MarkdownIt from 'markdown-it'
 import { warn } from '../cli'
+import { debug } from '../utils/debug'
+
+type Token = ReturnType<MarkdownIt['parse']>[number]
 
 export const keywordsAsArray = (keywords: unknown): string[] | undefined => {
   let kws: any[] | undefined
@@ -26,7 +30,41 @@ export const keywordsAsArray = (keywords: unknown): string[] | undefined => {
   return undefined
 }
 
-export default function metaPlugin({ marpit }: { marpit: Marpit }) {
+export const detectTitle = (tokens: Token[]): string | undefined => {
+  let bestHeading: { level: number; content?: string } = {
+    level: Number.MAX_SAFE_INTEGER,
+  }
+
+  const { length } = tokens
+
+  for (let i = 0; i < length; i += 1) {
+    const token = tokens[i]
+
+    if (token.type === 'heading_open' && !token.hidden) {
+      const tagMatcher = token.tag.match(/^h([1-6])$/i)
+
+      if (tagMatcher) {
+        const level = parseInt(tagMatcher[1], 10)
+        const likelyContentToken = tokens[i + 1]
+
+        if (
+          likelyContentToken &&
+          likelyContentToken.type === 'inline' &&
+          level < bestHeading.level
+        ) {
+          // TODO: Use the content of renderInline() as a title
+          bestHeading = { level, content: likelyContentToken.content.trim() }
+        }
+      }
+    }
+  }
+
+  return bestHeading.content
+}
+
+export default function metaPlugin(md: MarkdownIt & { marpit: Marpit }) {
+  const { marpit } = md
+
   Object.assign(marpit.customDirectives.global, {
     author: (v) => (typeof v === 'string' ? { marpCLIAuthor: v } : {}),
     description: (v) =>
@@ -38,6 +76,8 @@ export default function metaPlugin({ marpit }: { marpit: Marpit }) {
     },
     title: (v) => (typeof v === 'string' ? { marpCLITitle: v } : {}),
     url: (v) => {
+      if (Array.isArray(v)) return {}
+
       // URL validation
       try {
         if (v) new URL(v)
@@ -46,9 +86,36 @@ export default function metaPlugin({ marpit }: { marpit: Marpit }) {
         return {}
       }
 
-      return { marpCLIURL: v }
+      return { marpCLIURL: v ?? undefined }
     },
   })
 
-  // TODO: Add rule to fill meta from content of slide deck when directives are not defined.
+  md.core.ruler.after(
+    'marpit_directives_global_parse',
+    'marp_cli_meta_title_detection',
+    (state) => {
+      if (state.inlineMode) return false
+
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore Assign title global directive into protected property to allow reading title from other injected plugins.
+      const { lastGlobalDirectives } = (
+        state.md as MarkdownIt & { marpit: Marpit }
+      ).marpit
+
+      if (lastGlobalDirectives && !('marpCLITitle' in lastGlobalDirectives)) {
+        debug(
+          'Markdown parser had detected no title setting. Marp CLI will try to extract title from headings in Markdown contents.'
+        )
+
+        const detectedTitle = detectTitle(state.tokens)
+
+        if (detectedTitle) {
+          lastGlobalDirectives.marpCLITitle = detectedTitle
+          debug(`Title detected: "${detectedTitle}"`)
+        } else {
+          debug('No title was detected from headings.')
+        }
+      }
+    }
+  )
 }